Cast GetCurrentProcessId to int to avoid compiler warning.
git-svn-id: file:///fltk/svn/fltk/branches/branch-1.1@4277 ea41ed52-d2ee-0310-a9c1-e6b18d33e121

+++ fluid/undo.cxx
@@ -33,7 +33,7 @@
 #if defined(WIN32) && !defined(__CYGWIN__)
 #  include <io.h>
 #  include <windows.h>
-#  define getpid GetCurrentProcessId
+#  define getpid (int)GetCurrentProcessId
 #else
 #  include <unistd.h>
 #endif // WIN32 && !__CYGWIN__
